requirements.txt
adrf==0.1.6
asgiref==3.8.1
async-property==0.2.2
attrs==23.2.0
cachetools==5.4.0
certifi==2024.7.4
charset-normalizer==3.3.2
click==8.1.7
cloudinary==1.40.0
defusedxml==0.7.1
diff-match-patch==20230430
Django==5.0.7
django-autoslug==1.9.9
django-cloudinary-storage==0.3.0
django-cors-headers==4.4.0
django-debug-toolbar==4.4.6
django-import-export==4.0.0
django-jazzmin==3.0.0
django-querycount==0.8.3
djangorestframework==3.15.2
drf-spectacular==0.27.2
et-xmlfile==1.1.0
facebook-sdk==3.1.0
google-auth==2.32.0
h11==0.14.0
idna==3.7
inflection==0.5.1
iniconfig==2.0.0
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
MarkupPy==1.14
odfpy==1.4.1
openpyxl==3.1.5
packaging==24.1
pillow==10.4.0
pluggy==1.5.0
psycopg==3.2.1
pyasn1==0.6.0
pyasn1_modules==0.4.0
PyJWT==2.8.0
pytest==8.3.2
pytest-django==4.8.0
python-decouple==3.8
PyYAML==6.0.1
referencing==0.35.1
requests==2.32.3
rpds-py==0.19.1
rsa==4.9
six==1.16.0
sqlparse==0.5.1
tablib==3.5.0
typing_extensions==4.12.2
uritemplate==4.1.1
urllib3==2.2.2
uvicorn==0.30.3
whitenoise==6.7.0
xlrd==2.0.1
xlwt==1.3.0
